fix variant item price bug

+++ models/Variant.go
@@ -18,6 +18,6 @@ type VariantItem struct {
 	VariantID uuid.UUID `json:"variantId"`
 	Variant   Variant   `gorm:"foreignKey:VariantID;references:ID" json:"variant"`
 	Value     string    `gorm:"not null" json:"value"` // 100ml - blue - 6.5
-	Price     float64   `gorm:"default:0" json:"price"`
+	Price     int       `gorm:"default:0" json:"price"`
 	Quantity  int       `gorm:"default:0" json:"quantity"` //
 }
